FIX: Handle readonly / nomodifiable buffer warning / error
ENH: In all mappings, include all lines of closed folds and interpret [count] as the number of visible lines, i.e. count closed folds as one count. With this, moves and duplications are always done "over", never "into" closed folds, and the 'relativenumber' column can be used to determine the [count] to reach a target.